add: support browser
