+++ .gitignore
@@ -81,3 +81,4 @@ crashlytics.properties
 crashlytics-build.properties
 
 
+dist/
