+++ .vscode/settings.json
@@ -1,0 +1,5 @@
+{
+  "yaml.schemas": {
+    "https://www.artillery.io/schema.json": []
+  }
+}
